Fix interrupt mapping
Interrupt lines were incorrectly attached to the same lines that the push button was attached to, which was pulling them high whenever the user push button was pressed.  Move them to bank E.  Also had a problem where the system configuration peripheral was not enabled; this was preventing modifications to EXTI and NVIC.  Turn it on for the purpose of altering settings.

Also move the accelerometer interrupt to INT2.  Apparently the pullup did something funny to INT1 and I can't get it working anymore on this board.