Interrupt handlers must also consider the current line level
Sometimes the interrupt line is already asserted on device reset, and depending on bandwidth, it's possible that the read operation will never catch up and emtpy the queue thereby deasserting the interrupt line.  We need to ensure that this possible condition does not result in a device hang by checking the line level at the end of the handler and then repending, and also checking the line level right after initialization.